server/requirements.txt
fastapi
pydantic
pydantic-settings
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pymongo
google-generativeai
langchain
langchain-google-genai
pinecone-client
python-dotenv
psycopg2-binary
motor
python-docx
PyPDF2
scikit-learn
vaderSentiment
spacy
lucide-react
